[algorithms] Move gc helper algorithms to separate files.
Rename files and algorithms to match the naming pattern.
