+++ include/boost/geometry/algorithms/detail/gc_topological_dimension.hpp
@@ -6,8 +6,8 @@
 // Licensed under the Boost Software License version 1.0.
 // http://www.boost.org/users/license.html
 
-#ifndef BOOST_GEOMETRY_ALGORITHMS_DETAIL_GC_HPP
-#define BOOST_GEOMETRY_ALGORITHMS_DETAIL_GC_HPP
+#ifndef BOOST_GEOMETRY_ALGORITHMS_DETAIL_GC_TOPOLOGICAL_DIMENSION_HPP
+#define BOOST_GEOMETRY_ALGORITHMS_DETAIL_GC_TOPOLOGICAL_DIMENSION_HPP
 
 
 #include <algorithm>
@@ -48,4 +48,4 @@ inline int gc_topological_dimension(GeometryCollection const& geometry)
 
 }} // namespace boost::geometry
 
-#endif // BOOST_GEOMETRY_ALGORITHMS_DETAIL_GC_HPP
+#endif // BOOST_GEOMETRY_ALGORITHMS_DETAIL_GC_TOPOLOGICAL_DIMENSION_HPP
